changed division method to double type to avid rounding errors

+++ src/main/java/com/codeup/blog/controllers/MathController.java
@@ -46,7 +46,7 @@ public class MathController {
 //  DIVIDE:
     @GetMapping("/divide/{num1}/by/{num2}")
     @ResponseBody
-    public int divide(@PathVariable int num1, @PathVariable int num2) {
+    public double divide(@PathVariable double num1, @PathVariable double num2) {
         return num1 / num2;
     }
